+++ CHANGELOG.md
@@ -15,6 +15,9 @@ PLEASE REMEMBER TO CHANGE THE '..main' WITH AN ACTUAL TAG in GITHUB LINK.
   * `jnp.poly*` functions now require array-like inputs ({jax-issue}`#7732`)
   * `jnp.unique` and other set-like operations now require array-like inputs
     ({jax-issue}`#7662`)
+* New features:
+  * New debugging utility {func}`jax.debugging.make_hlo` to generate HLO or
+    optimized HLO for a JAX function ({jax-issue}`#7719`).
 
 ## jax 0.2.19 (Aug 12, 2021)
 * [GitHub